fetch-pack.c: use .string = .util for packfile_uri
Change code added in dd4b732df73 (upload-pack: send part of packfile
response as uri, 2020-06-10) to simply use the "util" member of the
"struct string_list", instead of storing a single string in the list
and needing to extract it out with careful use of "%.*s" etc. formats.

This allocates a bit more memory, but this code is never going to be
performance sensitive, so we should optimize this for readability and
to avoid the chance of making mistakes if we're off-by-one.

Since we have separate fields now we can also use the "string" to
append to "pack_lockfiles", instead of adding a '\0' to
"packname". This is right after a memcmp() that's limited to "hexsz",
so we know that the previous "packname" and our "item->string" have
the same contents.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>