pkt-line API: add API for deferred trace logging
Add an API to pkt-line that'll allow for a more general way to
accomplish what we recently did in 88e9b1e3fcb (fetch-pack: redact
packfile urls in traces, 2021-11-10), and without pkt-line.c itself
having to know about the structure of extensions in the v2 protocol.

The trace logging in the API happens at a fairly low-level, there's
both an OO-like interface and a functional interface to this API. Here
we're adding the bare minimum to support deferred tracing.

We now take an optional "struct strbuf *" argument in
packet_read_with_status(), as well as a new packet_reader_read_trace()
function that's like packet_reader_read(), but allows for passing this
"struct strbuf *" in.

Then at the lower packet_read_with_status() level we'll append the
trace we would otherwise have written to our trace target to this
"struct strbuf *". The user of the API is then expected to
change/amend/add to the trace, and call the now-public
packet_trace(). The "write" is something the user is expected to know
how to set (i.e. to know whether they invoked a "read" or "write"
function.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>