fetch-pack + pkt-line API: use deferred trace API for packfile-uri
Change the recently added code to redact packfile-uri traces to use
the deferred trace API added to pkt-line.[ch] in the preceding commit.

This amends code recently added in 88e9b1e3fcb (fetch-pack: redact
packfile urls in traces, 2021-11-10) to be driven from the part of
fetch-pack.c that's already tasked with parsing these packfile URIs, Instead of needing to have the the pkt-line API know about the
structure of the v2 protocol.

It's also a more narrow change, as we'll only defer and amend the
specific URI lines we get back, before we'd turn this on before
invoking process_section_header(). Before this the
find_packfile_uri_path() function would be asked to find URLs in
e.g. this sideband output:

    Total 2 (delta 0), reused 0 (delta 0), pack-reused 0

We'll also know whether or not we're processing a SHA-1 or SHA-256
dialog by doing this at the fetch-pack.c level, and can thus remove
the code that checked "len == 40 || len == 64" in the pre-image.

The traces look the same as before, i.e. without redaction:

    packet:        clone< \1packfile-uris
    packet:        clone< \131bccd8e911a271e162df220d1e718876e09447f http://127.0.0.1:5702/dumb/mypack-31bccd8e911a271e162df220d1e718876e09447f.pack
    packet:        clone< 0001

And with:

    packet:        clone< \1packfile-uris
    packet:        clone< \131bccd8e911a271e162df220d1e718876e09447f http://127.0.0.1:5702/<redacted>
    packet:        clone< 0001

The difference is that we'll only get that " \1131bccd..." line passed
to our find_packfile_uri_path(). See my [1] and [2] for the original
suggestion to take this approach.

1. https://lore.kernel.org/git/211109.86mtmedrhr.gmgdl@evledraar.gmail.com/
2. https://lore.kernel.org/git/211111.86tugjpn1x.gmgdl@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>