Makefile: have the TRACK_TEMPLATE printf, not echo
Fix a bug that's been with us for a long time, the recipes for the
various "TRACK" variables and files used "echo" to write out the file,
which will do its own interpolation:

    $ echo foo\\bar
    foo\bar

Thus providing a value that 'echo' would munge would fool the Makefile
into thinking that these variables had changed, which can be
reproduced e.g. with:

    make version.o GIT_USER_AGENT='foo"bar"\baz\'

On "master" this will keep re-building version.o, because we're fooled
into thinking that it changed, since we conflated our quoting. But
we'll still pass the properly quoted argument to the compiler via
-D.

This class bug has been with us ever since ca3bcabf118 (auto-detect
changed prefix and/or changed build flags, 2006-06-15), and since then
the recipe got copy/pasted for in making various other "TRACK"
files. For GIT-USER-AGENT it's been misquoting things since
620c293abdc (Makefile: split GIT_USER_AGENT from GIT-CFLAGS,
2012-06-20).

I.e. on "master" we'll do this:

    $ SHELL_PATH='sh -x' make version.o GIT_USER_AGENT='f\o\\o"\ba\\r"' 2>&1 | grep test.*ba; cat GIT-USER-AGENT
    + test xf\o\\o"\ba\\r" != xf\o\oa\r"
    f\o\oa\r"

Note that the quoting of the Makefile version of the variable and what
we write out to "GIT-USER-AGENT" is different. Now we'll instead write
the properly quoted and unescaped value to the file via "printf":

    $ SHELL_PATH='sh -x' make version.o GIT_USER_AGENT='f\o\\o"\ba\\r"' 2>&1 | grep test.*ba; cat GIT-USER-AGENT
    + test xf\o\\o"\ba\\r" != xf\o\\o"\ba\\r"
    f\o\\o"\ba\\r"

The trailing "\n" isn't strictly needed, and arguably makes this more
confusing, i.e. we'll add a "\n" purely for the convenience of
manually inspecting the file. It'll be stripped out with the "echo"
before by the shell. Let's retain it as before, only because manually
inspecting the files is a bit easier if they've got a trailing "\n".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>