Makefile: declare shell and C quoting in one place
Move the *_SQ, *_CQ and *_CQ_SQ variables that weren't declared along
with most of the rest of them to the big list such variables.

Since these are not simply expanded variables we can rely on their
interpolation being deferred, so it doesn't really matter that we move
them out of the relevant "ifdef" blocks, and as we'll see in a
subsequent commit declaring them in one place allows us to create them
all at once with a new "$(call)" function.

While we're at it let's rename the oddly lower-case
"procfs_executable_path_SQ" to "PROCFS_EXECUTABLE_PATH_SQ". The only
code that uses it is in this diff.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>