Makefile: properly quote C strings passed via -D<macro>[=<defn>]
Change the passing of "-D<macro>" variables to use the "shqq_Cq"
escaping function, and rename or migrate to *_SQ_CSQ suffixed
variables, for "shell quote, C quoted string".

Before this setting most of these to a path that contained a '"' or
'\' would error out on compilation, e.g. setting "mandir" to such a
path was one way to reproduce it, but this applied to almost all of
these:

    $ make -W builtin/help.c builtin/help.o mandir='/tmp/"foo"' V=1
    [...]
	builtin/help.c:437:35: error: expected ')'
        	char *git_man_path = system_path(GIT_MAN_PATH);
                	                         ^
	<command line>:26:29: note: expanded from here
	#define GIT_MAN_PATH "/tmp/"foo""
                            ^
	builtin/help.c:437:34: note: to match this '('
        	char *git_man_path = system_path(GIT_MAN_PATH);
                	                        ^
    [...]

Pretty much the ones that didn't have this issue were the
"DEFAULT_EDITOR" etc. variables which were added (or based on the
template of) 8f4b576ad14 (Provide a build time default-editor setting,
2009-10-30). This would work on "master", and still does:

    $ make editor.o DEFAULT_EDITOR='/tmp/"foo"\"bar"/emacs

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>