Makefile: split up and sort SCRIPT_DEFINES & TRACK_PREFIX
Re-format the SCRIPT_DEFINES and TRACK_PREFIX variables to be
incrementally defined instead of being defined with more than one
variable per-line, this makes subsequent diffs easier to read.

The values are no longer ":"-separated, but whitespace-separated,
that's OK to change, as we only care if the files change, not about
the specific contents of them.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>