CI: remove support for Azure CI
Remove various code required to support Azure CI. Recently in
4a6e4b96026 (CI: remove Travis CI support, 2021-11-23) we removed the
Travis CI support, which was last used in June 2021.

The supporting infrastructure for Azure CI was already partially
removed in 6081d3898fe (ci: retire the Azure Pipelines definition,
2020-04-11). As that commit notes:

    We have GitHub Actions now. Running the same builds and tests in Azure
    Pipelines would be redundant, and a waste of energy.

This follow-up removes the relevant Azure Pipelines-only supporting
code. Most of it was added in commits merged as part of
57cbc53d3e0 (Merge branch 'js/vsts-ci', 2019-02-06).

The motivation for this change is wanting to make some test-lib.sh
changes. If we don't remove the Azure-specific JUnit code changing
anything to do with writing the test TAP output will also need to
write out corresponding JUnit. I'd rather not figure out how to patch
dead code, so let's just remove it.

The "BREW_INSTALL_PACKAGES" variable being removed here was added in
b011fabd6e0 (ci/lib.sh: encapsulate Travis-specific things,
2019-01-27), but its last user went away in my 4a6e4b96026 (CI: remove
Travis CI support, 2021-11-23).

I had an RFC proposal to do some of this a while ago by generating the
JUnit via tap2junit[1]. Johannes had an objection to it on the basis
that if we resurrected the JUnit support in some form he'd prefer not
to run Perl on Windows[2] (which tap2junit is implemented in).

I have local WIP patches in C (via a new test-tool helper) to parse
the TAP output. That'll allow us to emit to any arbitrary targets,
including JUnit. So given that this is currently unused I'd prefer to
take this step first. The unsubmitted series to get that TAP parser
working would run into the same chicken & egg problem with needing to
concern itself with the JUnit output.

Note that the code being modified in "t/helper/test-run-command.c"
could also be entirely removed. It has been unused since
4c2c38e800f (ci: modification of main.yml to use cmake for vs-build
job, 2020-06-26). Before that in 6081d3898fe (ci: retire the Azure
Pipelines definition, 2020-04-11) the Azure Pipelines stopped using
it. Per the discussion starting at [3] there are some future plans to
make use of it again, so let's keep it for now.

1. https://lore.kernel.org/git/20210309160219.13779-8-avarab@gmail.com/
2. https://lore.kernel.org/git/nycvar.QRO.7.76.6.2103221429200.50@tvgsbejvaqbjf.bet
3. https://lore.kernel.org/git/patch-1.1-d1e464da0a9-20210906T002938Z-avarab@gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>