test-lib.sh: add a GIT_TEST_PASSING_SANITIZE_LEAK=check mode
Add a "GIT_TEST_PASSING_SANITIZE_LEAK=check" mode, which compliments
the existing "GIT_TEST_PASSING_SANITIZE_LEAK=true" as described in the
updated "t/README" file.

The new "--invert-exit-code" option needed to drive this isn't useful
on its own, normally we would not want to hide our failing tests.

But it is useful for these sorts of 1=1 marking "GIT_TEST_*" modes for
the reasons noted in "t/README".

We're not bothering to support "--write-junit-xml" in combination with
"--invert-exit-code", and will error out if they're combined. As noted
in [1] the former is dead code that should be removed, but until
that's done let's explicitly error when the two are combined.

1. https://lore.kernel.org/git/patch-1.1-eec0a8c3164-20211217T000418Z-avarab@gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>