object-file API: replace some use of check_object_signature()
Add a "hash_object_file_literally()" function to go with the existing
"hash_object_file()" function. This is currently a wrapper for its
sibling, but this change will allow us to change it to take an "enum
object_type" in a subsequent commit.

The only caller that wanted to ass a custom type to
"check_object_signature()" was the "git fsck" via its
"read_loose_object()", which is being changed here.

There was an existing hash_object_file_literally() which I'm renaming
to "hash_write_object_file_literally()", that function is only used
for "hash-object --literally". It's confusing to have a
"hash_object_file_literally()" and it and "hash_object_file()" that do
very different things. See 0c3db67cc81 (hash-object --literally: fix
buffer overrun with extra-long object type, 2015-05-04) for the
initial implementation of the "hash-object" function.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>