object-file API: replace check_object_signature() with stream_*
Change the task of the the check_object_signature() function so that
it no longer confusingly does two unrelated things.

Now the "!map" branch of it is exposed as a renamed
stream_object_signature(). In preceding commits we migrated all of the
"if (map)" callers over to "hash_object_file_oideq()" or
"hash_object_file_literally()" followed by "!oideq(...)".

This makes the caller in parse_object() less confusing, as we'll now
no longer need to pass NULL for "map" and "OBJ_BAD" for the unused
"type" argument. The "verify_packfile()" caller is arguably a bit
worse off than before, but as we're only using this function in a
couple of places now a bit more verbosity there is acceptable.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>