catfile: use argv, argc for --stdin-cmd command function signature
WIP: To allow for future commands that may take more than 1 argument, make
the signature of the fn in the parse_cmd struct include argc, argv. argc
is the number of arguments that were found in the line, and argv is a
NULL terminated array of strbufs that contain the arguments that were
found on the line.

DIE if the arguments found do not match the number specified in the
parse_cmd struct.

This change will allow the --stdin-cmd mode to support:

<command> NUL <arg1> NUL <arg2> NUL
<command> SP <arg1> SP <arg2> LF

With an arbitrary amount of arguments.

Signed-off-by: John Cai <johncai86@gmail.com>