grep.[ch]: use a "switch" for "enum grep_pattern_type"
Take advantage of the "grep_pattern_type" being an "enum" by using a
"switch" over its known fields in compile_regexp().

The compiler will now be able to help us if we'll add a new
"opt->pattern_type_option" in the future, which makes this code more
maintainable. Even without that we'll get the advantage of avoiding
repeated checking of "opt->pattern_type_option" within the function,
we'll instead check new "pcre2" and "posix" variables.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>