parse-options API: assert no "argh" or "help" on "hidden"
Extend parse_options_check() to check for whether "argh" or "help" are
non-NULL when the PARSE_OPT_HIDDEN flag is being used. This will
ensure that the sort of dead code that was removed in preceding
commits won't recur.

We need to exclude "git rev-parse --parseopt" from this check. Let's
add a new "PARSE_OPT_REV_PARSE_PARSEOPT" flag for it to parse in. When
it's in use we'll skip this stricter check.

We could amend rev-parse.c itself to stop populating "argh" and
"help", but doing so would require changing its stateful parsing, it
will xstrdup() the "help" before it gets to parsing the "flags". It's
possible to change it, but let's just treat it as a black box instead.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>