parse-options API: require { 0 }-initialized "ctx", drop wrapper
In the preceding commit we got rid of "OPT_ALIAS", and with the
removal of the preprocess_options() function call between the memset()
of "ctx" in parse_options() and parse_options_start_1() we're not
going to have "ctx" be zero'd out, then potentially modified in
preprocess_options(), only to be zero'd out again in
parse_options_start().

Let's instead migrate the few callers of parse_options_start() to "{ 0
}"-initialize their "struct parse_opt_ctx_t ctx", and get rid of the
parse_options_start() wrapper for parse_options_start_1(). We can now
call the old parse_options_start_1() directly.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>