tag.c: add a "taglen" variable to parse_tag_buffer()
Avoid repeated references to "nl - bufptr" and instead create a
"taglen" variable to use in the few places where "nl - taglen" will
always yield return the same number. As we later re-assign both
"bufptr" and "nl" in the body of this function creating this temporary
variable makes the code easier to follow.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>