*.c: amend unreachable "return" statements for SunCC
Amend or remove unreachable return statements added to the reftable
code in [1] and [2], and to "test-xml-encode.c" in [3]. This avoids
the following warnings on at least versions 12.5 and 12.6 of SunCC
from Oracle Developer Studio:

    "builtin/am.c", line 1829: warning: statement not reached
    "reftable/refname.c", line 135: warning: statement not reached
    "t/helper/test-xml-encode.c", line 79: warning: statement not reached

In the case of am_run(), validate_refname() and cmd__xml_encode() let's just
"break" from the loop into the previously unreachable "return 0" that
was there, instead of removing that return statement. This solution
was suggested by Johannes Sixt[4].

Curiously the:

    "reftable/writer.c", line 42: warning: statement not reached

Is only something that's emitted by an older sub-version of 12.6,
i.e. it happens on:

    avar@gcc-solaris11:~$ /opt/developerstudio12.6/bin/suncc -V
    cc: Studio 12.6 Sun C 5.15 SunOS_sparc 152881-01 2018/03/16

But not:

    avar@gcc-solaris10:~$ /opt/developerstudio12.6/bin/suncc -V
    cc: Studio 12.6 Sun C 5.15 SunOS_sparc 152881-04 2019/03/08

It might also have to do with the Solaris version (although I doubt
it), in any case let's fix that one like the rest.

1. acb533440fc (reftable: implement refname validation, 2021-10-07)
2. f14bd719349 (reftable: write reftable files, 2021-10-07)
3. 22231908151 (tests: optionally write results as JUnit-style .xml,
   2019-01-29)
4. https://lore.kernel.org/git/9ce034fd-b696-60d2-c292-98285aff180a@kdbg.org/