gpg tests: don't block on /dev/random by making a GPG key
Amend a test added in 1bfb57f642d (ssh signing: test that gpg fails
for unknown keys, 2021-09-10) to stop generating a new GPG key. As
noted in the discussion downthread of [1] generating a key would cause
this test to hang on some platforms until randomness could be
retrieved from /dev/random, which on some setups would be some
approximation of forever.

For this test we don't need to generate a key at all. This test is
meant to check the case of an unknown signature, so we can sign using
one of our existing GPG keys, and then set "GNUPGHOME" in the
environment to a directory set up to not have any keys.

1. https://lore.kernel.org/git/211222.86ilvhpbl0.gmgdl@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>